Add login program
This completes the login flow. qgetty execve's the login binary which prompts for a
password, verifies it with the previously created shadow library in libq, and then
execve's the shell from /etc/passwd

If the password is incorrect n times (n=5 at the moment), we crash out. Next step is to
have qinit create the getty processes so they're not pid 1, so crashing out just means
we restart it