Added history support in sh
This required a bit of a rewrite of things, particularly the input from the terminal.
Previously we were using the echo functionality of the terminal, but this doesn't let us
intercept the characters we want (Up, Down, etc). In this commit we rewrite this to turn off
echoing, and do it ourselves