Added "Courthouse" implementation to qinit
A "Courthouse" is a wrapper around an Inotify client which listens for given events for
a number of paths, and notifies a registry once it has seen all of them.

The purpose of this is to handle Start Conditions (e.g. Service A starts a unix socket),
- things which we expect to happen before the service is properly considered "started".
The driving service behind this is Freudian - Freudian creates a Unix Socket, and shouldn't be
considered started until that socket is open. If we start dependent services before the socket is open
we get a race condition where the dependant service might try and make a freudian request before the
socket is open